requirements.txt
astropy==4.0.1.post1
cffi==1.14.0
cryptography==2.9.2
cycler==0.10.0
dqsegdb2==1.0.1
gwdatafind==1.0.4
gwosc==0.5.3
gwpy==1.0.1
h5py==2.10.0
kiwisolver==1.2.0
ligo-segments==1.2.0
ligotimegps==2.0.1
matplotlib==3.2.2
numpy==1.19.0
piexif==1.1.3
Pillow==7.2.0
pkg-resources==0.0.0
pycparser==2.20
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.5.0
six==1.15.0
tqdm==4.47.0
